Include updates for PFTaus Version to tag w. validated electron R9 fixes

+++ interface/HLTJets.h
@@ -109,7 +109,7 @@ private:
     //PFTau
     int nohPFTau;
     float *ohpfTauEta,*ohpfTauPhi,*ohpfTauPt,*ohpfTauJetPt,*ohpfTauLeadTrackPt,*ohpfTauLeadPionPt;
-    int *ohpfTauTrkIso, *ohpfTauGammaIso;
+    float *ohpfTauTrkIso, *ohpfTauGammaIso;
     //PFJets
     float pfMHT;    
     int nohPFJet;
